+++ wazo_calld/exceptions.py
@@ -21,22 +21,6 @@ class CalldUninitializedError(APIException):
         )
 
 
-class WazoAmidError(APIException):
-
-    def __init__(self, wazo_amid_client, error):
-        super().__init__(
-            status_code=503,
-            message='wazo-amid request error',
-            error_id='wazo-amid-error',
-            details={
-                'wazo_amid_config': {'host': wazo_amid_client.host,
-                                     'port': wazo_amid_client.port,
-                                     'timeout': wazo_amid_client.timeout},
-                'original_error': str(error),
-            }
-        )
-
-
 class ARIUnreachable(Exception):
 
     def __init__(self, ari_config, original_error=None):
