+++ wazo_calld/plugins/transfers/stasis.py
@@ -10,8 +10,8 @@ from ari.exceptions import ARINotFound
 from ari.exceptions import ARINotInStasis
 
 from wazo_calld.ari_ import DEFAULT_APPLICATION_NAME
-from wazo_calld.exceptions import WazoAmidError
 from wazo_calld.plugin_helpers.ari_ import Channel
+from wazo_calld.plugin_helpers.exceptions import WazoAmidError
 
 from . import ari_helpers
 from .event import (
